requirements.txt
mysql-connector-python
spacy
beautifulsoup4
requests
json
